pyproject.toml
[project]
name = "hermes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio>=3.4.3",
    "feedparser>=6.0.11",
    "openai>=1.76.2",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "reportlab>=4.4.1",
    "slack-sdk>=3.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.2.0",
]
